refactor: remove some unused functions from McpUtils file and get rid of that file while moving a required function in McpUtils file to SpecmaticMcpServer
